+++ src/main/java/leaguehub/leaguehubbackend/exception/email/EmailExceptionHandler.java
@@ -1,0 +1,43 @@
+package leaguehub.leaguehubbackend.exception.email;
+
+import leaguehub.leaguehubbackend.exception.email.exception.DuplicateEmailException;
+import leaguehub.leaguehubbackend.exception.email.exception.InvalidEmailAddressException;
+import leaguehub.leaguehubbackend.exception.global.ExceptionCode;
+import leaguehub.leaguehubbackend.exception.global.ExceptionResponse;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+@Slf4j
+@RestControllerAdvice
+@RequiredArgsConstructor
+public class EmailExceptionHandler {
+
+    @ExceptionHandler(InvalidEmailAddressException.class)
+    public ResponseEntity<ExceptionResponse> invalidEmailAddress(
+            InvalidEmailAddressException e
+    ) {
+        ExceptionCode exceptionCode = e.getExceptionCode();
+        log.error("{}", exceptionCode.getMessage());
+
+        return new ResponseEntity<>(
+                new ExceptionResponse(exceptionCode),
+                exceptionCode.getHttpStatus()
+        );
+    }
+
+    @ExceptionHandler(DuplicateEmailException.class)
+    public ResponseEntity<ExceptionResponse> invalidEmailAddress(
+            DuplicateEmailException e
+    ) {
+        ExceptionCode exceptionCode = e.getExceptionCode();
+        log.error("{}", exceptionCode.getMessage());
+
+        return new ResponseEntity<>(
+                new ExceptionResponse(exceptionCode),
+                exceptionCode.getHttpStatus()
+        );
+    }
+}
