+++ src/main/java/leaguehub/leaguehubbackend/exception/email/exception/DuplicateEmailException.java
@@ -1,12 +1,14 @@
-package leaguehub.leaguehubbackend.exception.member.exception;
+package leaguehub.leaguehubbackend.exception.email.exception;
 
+import leaguehub.leaguehubbackend.exception.email.EmailExceptionCode;
 import leaguehub.leaguehubbackend.exception.global.ExceptionCode;
 
-import static leaguehub.leaguehubbackend.exception.member.MemberExceptionCode.DUPLICATE_EMAIL_EXCEPTION;
+import static leaguehub.leaguehubbackend.exception.email.EmailExceptionCode.DUPLICATE_EMAIL_EXCEPTION;
+
 
 public class DuplicateEmailException extends RuntimeException{
 
-    private final ExceptionCode exceptionCode;
+    private final EmailExceptionCode exceptionCode;
 
     public DuplicateEmailException() {
         super(DUPLICATE_EMAIL_EXCEPTION.getMessage());
