+++ src/main/java/leaguehub/leaguehubbackend/exception/member/MemberExceptionHandler.java
@@ -2,8 +2,6 @@ package leaguehub.leaguehubbackend.exception.member;
 
 import leaguehub.leaguehubbackend.exception.global.ExceptionCode;
 import leaguehub.leaguehubbackend.exception.global.ExceptionResponse;
-import leaguehub.leaguehubbackend.exception.member.exception.DuplicateEmailException;
-import leaguehub.leaguehubbackend.exception.member.exception.InvalidEmailAddressException;
 import leaguehub.leaguehubbackend.exception.member.exception.MemberNotFoundException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -29,29 +27,5 @@ public class MemberExceptionHandler {
         );
     }
 
-    @ExceptionHandler(InvalidEmailAddressException.class)
-    public ResponseEntity<ExceptionResponse> invalidEmailAddress(
-            InvalidEmailAddressException e
-    ) {
-        ExceptionCode exceptionCode = e.getExceptionCode();
-        log.error("{}", exceptionCode.getMessage());
 
-        return new ResponseEntity<>(
-                new ExceptionResponse(exceptionCode),
-                exceptionCode.getHttpStatus()
-        );
-    }
-
-    @ExceptionHandler(DuplicateEmailException.class)
-    public ResponseEntity<ExceptionResponse> invalidEmailAddress(
-            DuplicateEmailException e
-    ) {
-        ExceptionCode exceptionCode = e.getExceptionCode();
-        log.error("{}", exceptionCode.getMessage());
-
-        return new ResponseEntity<>(
-                new ExceptionResponse(exceptionCode),
-                exceptionCode.getHttpStatus()
-        );
-    }
 }
